+++ core/src/main/java/org/jdbi/v3/core/enums/DatabaseValue.java
@@ -17,6 +17,7 @@ import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
+
 import org.jdbi.v3.core.qualifier.Qualifier;
 import org.jdbi.v3.meta.Beta;