+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@
   - `ColumnMappers.setNullPrimitivesToDefaults(boolean)` allows you to decide if database nulls
     should become Java primitive defaults or a mapping exception.
   - `@PropagateNull` annotation allows a missing property to indicate a missing compound value
+  - `@DatabaseValue` annotation lets you customize enum values as stored in the DB
 - Bug Fixes
   - Immutables integration doesn't respect @Value.Default for primitives that are nulled in the db
 - Improvements